% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripe-class.R
\name{belongs}
\alias{belongs}
\alias{belongs,ANY,ANY-method}
\alias{belongs,numeric,Stripe-method}
\alias{belongs,Stripe,Stripe-method}
\title{Check if one object belongs to an stripe}
\usage{
belongs(x, y, ...)

\S4method{belongs}{numeric,Stripe}(x, y, strict = FALSE)

\S4method{belongs}{Stripe,Stripe}(x, y, strict = TRUE)
}
\arguments{
\item{x}{A numeric vector or a stripe}

\item{y}{A stripe}

\item{strict}{A boolean vector of length one or two. Value
\code{TRUE} indicates strict inequality. Value \code{FALSE}
indicates weak inequality. If a vector of length two is passed,
the first element will be interpreted as the comparison to be
applied to the lower bound.}
}
\value{
A boolean vector with value TRUE is
}
\description{
Check if one object belongs to an stripe
}
