% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts-definitions.R, R/rulesets.R
\docType{methods}
\name{build_zones}
\alias{build_zones}
\alias{build_zones,xbarchart-method}
\alias{build_zones,ANY,ANY-method}
\alias{build_zones,cchart-method}
\alias{build_zones,pchart-method}
\alias{western_electric}
\title{The generic for \code{build_zones}}
\usage{
build_zones(object, ...)

\S4method{build_zones}{xbarchart}(object)

\S4method{build_zones}{cchart}(object)

\S4method{build_zones}{pchart}(object)

western_electric(object, rule = 1:4, ...)
}
\arguments{
\item{object}{An instance of \code{SpcData}}

\item{...}{Additional parameters. Currently not implemented}

\item{rule}{A vector of integers indicating the rule to be applied
to the data. It defaults to rules 1 to 4.}
}
\value{
An instance of \code{SpcData}

A data.frame of booleans with as many rows as data points
    in \code{data} and as many columns as rules specified in
    \code{rule}.
}
\description{
This generic allows for the creation of different methods to
generate the zones needed for each type of chart. Each chart is
expected to provide a \code{build_zones} method that takes the
\code{data} slot (see \code{spcchart}) and produces an instance of
\code{SpcData}.

This generic applies the Western Electric rules to \code{SpcData}
}
\seealso{
\code{\link{spcchart}} and \code{\link{spcdata}}
}
