% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runners.R
\name{chunkify}
\alias{chunkify}
\title{Applies a runner function to chunks of a given size}
\usage{
chunkify(x, runner, size)
}
\arguments{
\item{x}{A numeric vector}

\item{runner}{A function that returns a boolean}

\item{size}{The chunk size from \code{x} to be passed to \code{runner}}
}
\description{
Applies a runner function to chunks of a given size
}
\examples{
chunkify(c(1, 0, 1, 0), zigzag, 4)
}
