% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runners.R
\name{inside}
\alias{inside}
\title{Tests if a number of observations fall inside an interval}
\usage{
inside(x, howmany, interval)
}
\arguments{
\item{x}{A numeric vector}

\item{howmany}{An integer with how many observations in x need to
be inside an interval}

\item{interval}{A stripe}
}
\value{
A boolean with value TRUE if \code{howmany} observations in
    \code{x} fall inside of \code{interval}
}
\description{
Tests if a number of observations fall inside an interval
}
\examples{
inside(c(10, 10, 10, 10), 3, stripe(-100, 100))
}
