% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runners.R
\name{monotone}
\alias{monotone}
\title{Tests if a vector shows a monotonically increasing or decreasing pattern}
\usage{
monotone(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A boolean with value TRUE if x contains a monotonically increasing or decreasing
}
\description{
Tests if a vector shows a monotonically increasing or decreasing pattern
}
\examples{
monotone(c(1, 2, 3))
monotone(c(3, 2, 1))
}
