% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rulesets.R
\name{nelson,SpcData-method}
\alias{nelson,SpcData-method}
\alias{nelson,ANY,ANY-method}
\title{Methods for Western Electric ruleset}
\usage{
\S4method{nelson}{SpcData}(object, rule = 1:8, ...)
}
\description{
Methods for Western Electric ruleset
}
