% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runners.R
\name{outside}
\alias{outside}
\title{Test if a number of observations fall outside an interval}
\usage{
outside(x, howmany, interval, mixing = FALSE)
}
\arguments{
\item{x}{A numeric vector}

\item{howmany}{An integer with how many observations in x need to
be outside an interval}

\item{interval}{A stripe}

\item{mixing}{A boolean. If TRUE, outside observations need to be
on different sides of the interval. If TRUE, outside
observations need to be on the same side of the interval.
Default is FALSE}
}
\value{
A boolean with value TRUE if \code{howmany} observations in
    \code{x} fall outside of \code{interval}
}
\description{
Test if a number of observations fall outside an interval
}
\examples{
outside(c(10, 10, 10, 10), 3, stripe(-1, 1))
outside(c(10, 10, 10, 10), 3, stripe(-1, 1), mixing=TRUE)
}
