% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{runner}
\alias{runner}
\title{Basic runner}
\usage{
runner(x, howmany, outof, interval)
}
\arguments{
\item{x}{A numeric vector}

\item{howmany}{An integer}

\item{outof}{An integer}

\item{interval}{A stripe}
}
\value{
A vector of booleans
}
\description{
Basic runner
}
\examples{
runner(c(1, 2, 3, 4), 2, 3, rep(stripe(2, 2), 3))
}
