% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone-class.R
\docType{methods}
\name{zone}
\alias{zone}
\alias{zone,Zone-method}
\alias{setters-getters}
\alias{zone<-,Zone-method}
\title{The generic to get each zone from a \code{Zone} object}
\usage{
zone(object, region, ...)

\S4method{zone}{Zone}(object, region, ...)

\S4method{zone}{Zone}(object, region, ...) <- value
}
\arguments{
\item{object}{An instance of \code{Zone}}

\item{region}{One of "0" (for the centerline), "A", "B", "C" to
retrieve the stripe correspoding to that zone}

\item{value}{An instance of \code{Stripe}}
}
\value{
A stripe
}
\description{
A getter to retrieve the zone from a \code{Zone} object

A setter to change the value of the zone in a \code{Zone} object
}
