% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplespc-class.R
\docType{class}
\name{SimpleSpc-class}
\alias{SimpleSpc-class}
\title{The \code{SimpleSpc} class}
\description{
The \code{SimpleSpc} class
}
\section{Slots}{

\describe{
\item{\code{data}}{An instance of \code{SpcData}}

\item{\code{anomalies}}{A data.frame of booleans}
}}

