% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplespc-class.R, R/simplespc.R
\docType{package}
\name{simplespc}
\alias{simplespc}
\alias{simplespc-package}
\title{Main call to apply SPC to some data}
\usage{
simplespc(data, chart, ruleset = "WE", rules = 1:4)
}
\arguments{
\item{data}{Either a numeric vector/matrix or an instance of
\code{SpcData}}

\item{chart}{A string with the name of a chart. The chart is
expected to exist as a class inheriting from \code{SpcChart}.
See \code{\link{spcchart}}.}

\item{ruleset}{A string with the name of a given ruleset. Currently
only "WE" (for "Western Electric") is implemented.}

\item{rules}{An integer vector with the rules to be calculated. It
defaults to all rules for Western Electric.}
}
\value{
An instance of \code{SimpleSpc} that contains both the
    input data as well as a matrix of booleans indicating the
    anomalies corresponding to each rule.
}
\description{
This function is the main entry point for the package. It provides
a simple way to apply different rules for different types of data.
It is possible to use the function in training mode so that rather
than returning the anomalies to some given input data, it only
calculates the information necessary (an instance of
\code{SpcData}) that can be used to evaluate other input data.
}
\section{Contents}{

The package contains utilities to implement SPC charts
}

\author{
Gonzalo Rivero
}
