% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts-class.R
\docType{class}
\name{SpcChart-class}
\alias{SpcChart-class}
\alias{xbarchart-class}
\alias{cchart-class}
\alias{pchart-class}
\title{A base class for SPC models}
\description{
All charts are expected to inherit from this class. The class
defines a single argument \code{data} that can be either a numeric
vector or a matrix. The class does not impose any restrictions on
the matrix but it is intended to be used for charts that require
(observations, sample sizes)
}
\section{Slots}{

\describe{
\item{\code{data}}{An instance of numeric or matrix}
}}

