% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts-class.R
\docType{class}
\name{SpcData-class}
\alias{SpcData-class}
\title{An S4 class to represent SPC Data}
\description{
This class is used to host the necessary data to apply a ruleset to
given input data. The class facilitates building and testing the
validity of the different zones that define statistical process
control.
}
\section{Slots}{

\describe{
\item{\code{data}}{A numeric vector or a matrix. This slot should contain
the raw data that was passed to the chart.}

\item{\code{yhat}}{A numeric representation of the \code{data} slot that is
used as centerline.}

\item{\code{statistics}}{A numeric vector with the calculated location.}

\item{\code{deviation}}{A numeric vector with the standard deviation used
for the computation of the difference zones. It can have length
> 1 if the chart has variable limits.}

\item{\code{zones}}{A \code{zone} object with the definition of the three
zones ("0", "A", "B", "C") that are used to evaluate the input
data.}
}}

