% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripe-class.R
\name{length,Stripe-method}
\alias{length,Stripe-method}
\alias{length,ANY,ANY-method}
\alias{rep,Stripe-method}
\alias{rep,ANY,ANY-method}
\title{Calculate the length of a stripe}
\usage{
\S4method{length}{Stripe}(x)

\S4method{rep}{Stripe}(x, times)
}
\arguments{
\item{x}{A stripe}

\item{times}{An integer giving the non-negative number of times to
extend the stripe}
}
\value{
An integer

A stripe extended \code{times} times
}
\description{
Calculate the length of a stripe

Extend a stripe a number of times
}
\examples{
st <- stripe(c(1, 1), c(2, 2))
length(st)
st <- stripe(c(1, 1), c(2, 2))
rep(st, 2)
}
