% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripe-class.R
\docType{class}
\name{Stripe-class}
\alias{Stripe-class}
\alias{stripe}
\title{An S4 class to represent an stripe}
\usage{
stripe(lower = -Inf, upper = Inf)
}
\arguments{
\item{lower}{A numeric vector}

\item{upper}{A numeric vector}
}
\description{
This function creates a new instance of \code{Stripe} which is
defined by lower and upper bounds
}
\section{Slots}{

\describe{
\item{\code{lower}}{A numeric vector representing the lower bound of the stripe}

\item{\code{upper}}{A numeric vector representing the upper bound of the stripe}
}}

\examples{
stripe(c(1, 1), c(2, 2)) 
}
