% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rulesets.R
\name{western_electric,SpcData-method}
\alias{western_electric,SpcData-method}
\alias{western_electric,ANY,ANY-method}
\title{Methods for Western Electric ruleset}
\usage{
\S4method{western_electric}{SpcData}(object, rule = 1:4, ...)
}
\description{
Methods for Western Electric ruleset
}
