% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runners.R
\name{zigzag}
\alias{zigzag}
\title{Tests if a vector has an oscillating pattern}
\usage{
zigzag(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A boolean with value TRUE if x contains observations that
    oscillate
}
\description{
Tests if a vector has an oscillating pattern
}
\examples{
zigzag(c(1, 0, 1, 0))
}
