% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone-class.R
\docType{class}
\name{Zone-class}
\alias{Zone-class}
\title{An S4 class to represent zones for SPC}
\description{
The core idea behind SPC is the definition of zones/regions around
the centerline that are used to decide whether a given point is to
be considered an anomaly or not. This class allows the user to
manually define each of those zones. The user is also expected to
pass a centerline which is defined here as an instance of
\code{Stripe} in which upper and lower bounds are equal.
}
\section{Slots}{

\describe{
\item{\code{centerline}}{A stripe containing the definition centerline
(lower and upper bounds are equal)}

\item{\code{zoneA}}{A stripe with the definition of zone A}

\item{\code{zoneB}}{A stripe with the definition of zone B}

\item{\code{zoneC}}{A stripe with the definition of zone C}
}}

